refactor(server): split annotation wiring and conflict state out of mcp/
ADR-034 Unit 7a, part 2. Two more pieces that sat in `mcp/` only because
that is where they were written.

`wireAnnotationStore` depends on nothing else file-opener holds, and both
the open pipeline and `renameDocument` call it -- which is why
document-service reached for it through a dynamic `import("./file-opener")`
whose only job was breaking a cycle. Moved below both, that import becomes
static and the cycle is gone rather than deferred. Its `vi.spyOn` target
moves with it; the `toHaveBeenCalled` assertions added in #1643 are what
prove the spy still intercepts document-service's binding rather than
silently patching a namespace nobody reads.

`readPendingConflict` (document-service) and `flagExternalConflict`
(file-opener) are two halves of one piece of state -- `Y_MAP_EXTERNAL_CONFLICT`
-- that had no shared home. That split is why the file watcher had to
import document-service to ask a question about a map it writes itself.
Neither had an importer outside those two files, so this is a retarget,
not a re-export: no facade is left behind. Resolution stays in file-opener,
since clearing the flag runs the reload lifecycle.

Boundary tables re-derived from source, not hand-edited.
